refactor: Restructure zorphy_annotation with modular architecture
Refactored zorphy_annotation package into a clean, modular structure
with comprehensive documentation while maintaining 100% backward compatibility.

New Structure:
- annotations/ - Zorphy, Zorphy2, ZorphyX classes (4 files)
- core/ - Patch, PatchBase, EnhancedList (4 files)
- types/ - Opt<T> optional wrapper (2 files)
- utils/ - Generic JSON helpers (2 files)

Key Improvements:
- Clear separation of concerns across 4 modules
- Renamed List_E to EnhancedList (more descriptive)
- Added comprehensive documentation with usage examples
- Better discoverability and maintainability
- Zero breaking changes to existing API

Validation:
- ✅ dart analyze: 0 errors
- ✅ dart test: All tests passed
- ✅ Works with zorphy build_runner
- ✅ All existing code continues to work

File Changes:
- Created: 13 new files (~1,100+ lines)
- Deleted: 6 old files
- Modified: 2 files (zorphy.dart, zorphy_annotation.dart)

The annotation package now has a professional structure matching the
quality of the zorphy refactoring.
