+++ zorphy_annotation/lib/src/core/enhanced_list.dart
@@ -1,0 +1,150 @@
+/// Enhanced List type with additional functionality
+///
+/// This class wraps a standard [List] and provides the same interface
+/// with enhanced factory constructors and methods. It's designed to be
+/// a drop-in replacement for List in most contexts.
+///
+/// The name "EnhancedList" reflects its purpose better than the
+/// original "List_E" name.
+///
+/// ## Usage
+///
+/// ```dart
+/// final list = EnhancedList<String>.from(['a', 'b', 'c']);
+/// final transformed = list.map((s) => s.toUpperCase());
+/// final filtered = list.where((s) => s.startsWith('a'));
+/// ```
+///
+/// ## Factory Constructors
+///
+/// - `EnhancedList.empty()` - Creates an empty list
+/// - `EnhancedList.filled(length, fill)` - Creates fixed-length list
+/// - `EnhancedList.growable([length])` - Creates growable list
+/// - `EnhancedList.from(elements)` - Creates from iterable
+/// - `EnhancedList.generate(length, generator)` - Generates elements
+///
+/// ## Note
+///
+/// This class maintains compatibility with the original List_E class
+/// while providing better naming and documentation.
+class EnhancedList<E> {
+  final List<E> _list;
+
+  const EnhancedList(this._list);
+
+  /// Gets the underlying list
+  List<E> get list => _list;
+
+  /// Creates an empty EnhancedList
+  factory EnhancedList.empty() => EnhancedList(<E>[]);
+
+  /// Creates a fixed-length list with the given [fill] value
+  ///
+  /// All elements in the list will be initialized to [fill].
+  factory EnhancedList.filled(int length, E fill) =>
+      EnhancedList(List.filled(length, fill));
+
+  /// Creates a growable list with optional [length]
+  ///
+  /// If [length] is 0 (default), creates an empty growable list.
+  /// Otherwise, creates a list of specified length filled with null.
+  factory EnhancedList.growable([int length = 0]) {
+    if (length == 0) {
+      return EnhancedList(<E>[]);
+    }
+    return EnhancedList(List.filled(length, null as E));
+  }
+
+  /// Creates a list from an [Iterable]
+  factory EnhancedList.from(Iterable<E> elements) =>
+      EnhancedList(List.from(elements));
+
+  /// Creates a list with [length] positions using [generator]
+  ///
+  /// Each element is generated by calling [generator] with its index.
+  factory EnhancedList.generate(
+    int length,
+    E Function(int index) generator,
+  ) => EnhancedList(List.generate(length, generator));
+
+  /// Returns the number of elements in the list
+  int get length => _list.length;
+
+  /// Whether the list has no elements
+  bool get isEmpty => _list.isEmpty;
+
+  /// Whether the list has at least one element
+  bool get isNotEmpty => _list.isNotEmpty;
+
+  /// Gets the element at [index]
+  E operator [](int index) => _list[index];
+
+  /// Sets the element at [index] to [value]
+  void operator []=(int index, E value) {
+    _list[index] = value;
+  }
+
+  /// Adds [value] to the end of the list
+  void add(E value) => _list.add(value);
+
+  /// Adds all elements of [iterable] to the end of the list
+  void addAll(Iterable<E> iterable) => _list.addAll(iterable);
+
+  /// Removes [value] from the list
+  ///
+  /// Returns true if the value was in the list and was removed.
+  bool remove(Object? value) => _list.remove(value);
+
+  /// Removes all elements from the list
+  void clear() => _list.clear();
+
+  /// Returns an iterator for the list
+  Iterator<E> get iterator => _list.iterator;
+
+  /// Maps each element to a new value
+  ///
+  /// Returns a new EnhancedList with the mapped values.
+  EnhancedList<R> map<R>(R Function(E e) f) =>
+      EnhancedList(_list.map(f).toList());
+
+  /// Filters elements by [predicate]
+  ///
+  /// Returns a new EnhancedList with only elements satisfying [predicate].
+  EnhancedList<E> where(bool Function(E e) f) =>
+      EnhancedList(_list.where(f).toList());
+
+  /// Returns the first [n] elements
+  EnhancedList<E> take(int n) => EnhancedList(_list.take(n).toList());
+
+  /// Skips the first [n] elements
+  EnhancedList<E> skip(int n) => EnhancedList(_list.skip(n).toList());
+
+  /// Inserts all elements of [iterable] at [index]
+  void insertAll(int index, Iterable<E> iterable) =>
+      _list.insertAll(index, iterable);
+
+  /// Removes a range of elements from [start] to [end]
+  void removeRange(int start, int end) => _list.removeRange(start, end);
+
+  /// Sorts the list using optional [compare] function
+  void sort([int Function(E a, E b)? compare]) => _list.sort(compare);
+
+  @override
+  String toString() => _list.toString();
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is EnhancedList &&
+          runtimeType == other.runtimeType &&
+          _list.length == other._list.length &&
+          _list.every(
+            (element) => other._list[_list.indexOf(element)] == element,
+          );
+
+  @override
+  int get hashCode => _list.hashCode;
+}
+
+// Type alias for backward compatibility
+typedef List_E<E> = EnhancedList<E>;
