+++ zorphy_annotation/lib/src/src.dart
@@ -1,0 +1,12 @@
+// Barrel export for src directory
+export 'annotations/annotations.dart';
+export 'core/core.dart';
+export 'types/types.dart';
+export 'utils/utils.dart';
+
+// Re-export for backward compatibility
+export 'annotations/zorphy.dart';
+export 'annotations/zorphy2.dart';
+
+// Export List_E as type alias for backward compatibility
+export 'core/enhanced_list.dart' show List_E;
