+++ zorphy_annotation/lib/src/utils/utils.dart
@@ -1,0 +1,2 @@
+// Utility exports
+export 'generic_serialization.dart';
