+++ zorphy_annotation/lib/zorphy.dart
@@ -1,7 +1,4 @@
-export 'src/annotations.dart';
-export 'src/opt.dart';
-export 'src/List_E.dart';
-export 'src/patch.dart';
-export 'package:quiver/core.dart' show hashObjects;
-export 'package:json_annotation/json_annotation.dart';
-export 'src/patch_base.dart';
+// Export for backward compatibility
+export 'src/src.dart';
+export 'src/core/patch.dart';
+export 'src/core/patch_base.dart';
